pyproject.toml
[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.9.0",
    "gradio>=5.28.0",
    "matplotlib>=3.4.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "plotly>=6.0.1",
    "python-dotenv>=0.19.0",
    "requests>=2.25.0",
    "scikit-learn>=0.24.0",
    "scipy>=1.7.0",
]
